Add files via upload
In statistics, linear regression is a linear approach for modeling the relationship between a scalar response and one or more explanatory variables. 
The dataset is related to life expectancy with many factors in the 4 significant categories being health, social, economic, mortality, and immunization. We have to predict next year's value of average life expectancy using linear regression.